More Windows Fixes

+++ image_transport/include/image_transport/simple_publisher_plugin.h
@@ -68,12 +68,13 @@ template<class M>
 class SimplePublisherPlugin : public PublisherPlugin
 {
 public:
+ 
   virtual ~SimplePublisherPlugin() {}
-
+  
   virtual uint32_t getNumSubscribers() const
   {
     // TODO(mjcarroll) replace with publisher-specific call.
-    if (simple_impl_) {return simple_impl_->node_->count_subscribers(getTopic());}
+    if (simple_impl_) {return static_cast<uint32_t>(simple_impl_->node_->count_subscribers(getTopic()));}
     return 0;
   }
